// src/FrontEnd/warstar-frontend/src/Models/GridComponentType.ts
export interface GridComponentType {
    id: string; 
    type: string;
    name: string;
    shortName: string;
    selected: boolean;
    armour: number;
    structuralIntegrity: number;
    lifeSupport: number;
    minPowerDraw: number;
    maxPowerDraw: number;
    mass: number;
    price: number;
}